#24: Rename DatasetCreateOrSelect to DatasetViewer
