+++ app/src/pages/nanomine/xmlUploader/xmlUploader.html
@@ -78,8 +78,8 @@
     <div class="md-layout-item md-size-100">
       <h3>Inputs</h3>
       <div>
-        <dataset-create-or-select selectHeader="Choose or create a dataset" :selectedHandler="datasetSelectedHandler"
-          :datasetOptions="datasetOptions"></dataset-create-or-select>
+        <dataset-viewer selectHeader="Choose or create a dataset" :selectedHandler="datasetSelectedHandler"
+          :datasetOptions="datasetOptions"></dataset-viewer>
       </div>
       <div>
         <p>Select a completed Excel Template File
